requirements.txt
# API and Data Handling
google-api-python-client==2.88.0  # Required for YouTube Data API v3 handling
python-dotenv==1.0.0  # Required for loading environment variables
pandas==2.0.0  # Required for data manipulation and analysis
numpy==1.24.3  # Required for numerical operations
isodate==0.6.1  # Required for date parsing

# Visualization
matplotlib==3.7.1  # Required for creating static, interactive, and animated visualizations
seaborn==0.12.2  # Required for creating statistical data visualizations
plotly==5.14.1  # Required for creating interactive web-based visualizations
tabulate==0.9.0  # Required for creating tables


# Machine Learning and NLP
scikit-learn==1.2.2  # Required for machine learning algorithms and tools
nltk==3.8.1  # Required for natural language processing
vaderSentiment==3.3.2  # Required for sentiment analysis
sentence-transformers==2.2.2  # Required for semantic similarity analysis
# Dashboard
streamlit==1.22.0  # Required for creating interactive web applications (for visualization) 

# Data Processing
python-dateutil==2.8.2  # Required for parsing and manipulating dates and times
pytz==2023.3  # Required for time zone handling